+++ velox/experimental/stateful/state/HeapKeyedStateBackend.h
@@ -33,8 +33,8 @@ class HeapKeyedStateBackend : public KeyedStateBackend {
   std::shared_ptr<ValueState<uint32_t, int64_t, RowVectorPtr>>
   getOrCreateValueState(StateDescriptor& stateDescriptor) override;
 
-  std::shared_ptr<InternalTimerService<uint32_t, int64_t>> createTimerService(
-      Triggerable<uint32_t, int64_t>* triggerable) override;
+  std::shared_ptr<InternalTimerService<int64_t, int64_t>> createTimerService(
+      Triggerable<int64_t, int64_t>* triggerable) override;
 
   std::shared_ptr<ValueState<uint32_t, TimeWindow, RowVectorPtr>>
   getOrCreateGroupValueState(StateDescriptor& stateDescriptor) override;
@@ -45,9 +45,9 @@ class HeapKeyedStateBackend : public KeyedStateBackend {
   std::shared_ptr<MapState<uint32_t, int, uint32_t, RowVectorPtr>>
   getOrCreateRankMapState(StateDescriptor& stateDescriptor) override;
 
-  virtual std::shared_ptr<InternalTimerService<uint32_t, TimeWindow>>
+  virtual std::shared_ptr<InternalTimerService<int64_t, TimeWindow>>
   createGroupWindowAggTimerService(
-      Triggerable<uint32_t, TimeWindow>* triggerable) override;
+      Triggerable<int64_t, TimeWindow>* triggerable) override;
 
   void snapshot(
       int64_t checkpointId,
